functionseq api driven
