+++ src/plugins/functionseq/src/FunctionAPI.ts
@@ -1,0 +1,5 @@
+export class FunctionAPI {
+    emitMidiEvent(eventTime: number, time: number, data: number[]) {
+        
+    }
+}
